# Speedup DistinctValue collector and exit early for ingesters