--- a/modules/querier/querier.go
+++ b/modules/querier/querier.go
@@ -543,16 +543,13 @@ func (q *Querier) SearchTags(ctx context.Context, req *tempopb.SearchTagsRequest
 		return nil, fmt.Errorf("error querying ingesters in Querier.SearchTags: %w", err)
 	}
 
-	stopEarly := false
+outerLoop:
 	for _, resp := range lookupResults {
 		for _, res := range resp.response.(*tempopb.SearchTagsResponse).TagNames {
 			distinctValues.Collect(res)
 			if distinctValues.Exceeded() {
-				stopEarly = true
+				break outerLoop // break out of all loops
 			}
-		}
-		if stopEarly {
-			break
 		}
 	}
 
@@ -584,22 +581,15 @@ func (q *Querier) SearchTagsV2(ctx context.Context, req *tempopb.SearchTagsReque
 	limit := q.limits.MaxBytesPerTagValuesQuery(userID)
 	distinctValues := collector.NewScopedDistinctString(limit)
 
-	stopEarly := false
+outerLoop:
 	for _, resp := range lookupResults {
 		for _, res := range resp.response.(*tempopb.SearchTagsV2Response).Scopes {
 			for _, tag := range res.Tags {
 				distinctValues.Collect(res.Name, tag)
 				if distinctValues.Exceeded() {
-					stopEarly = true
-					break
+					break outerLoop // break out of all loops
 				}
 			}
-			if stopEarly {
-				break
-			}
-		}
-		if stopEarly {
-			break
 		}
 	}
 
@@ -632,7 +622,7 @@ func (q *Querier) SearchTagValues(ctx context.Context, req *tempopb.SearchTagVal
 
 	// Virtual tags values. Get these first.
 	for _, v := range search.GetVirtualTagValues(req.TagName) {
-		// virtual tags are small so no need to stopEarly here
+		// virtual tags are small so no need to stop early here
 		distinctValues.Collect(v)
 	}
 
@@ -643,17 +633,13 @@ func (q *Querier) SearchTagValues(ctx context.Context, req *tempopb.SearchTagVal
 		return nil, fmt.Errorf("error querying ingesters in Querier.SearchTagValues: %w", err)
 	}
 
-	stopEarly := false
+outerLoop:
 	for _, resp := range lookupResults {
 		for _, res := range resp.response.(*tempopb.SearchTagValuesResponse).TagValues {
 			distinctValues.Collect(res)
 			if distinctValues.Exceeded() {
-				stopEarly = true
-				break
+				break outerLoop // break out of all loops
 			}
-		}
-		if stopEarly {
-			break
 		}
 	}
 
@@ -698,16 +684,12 @@ func (q *Querier) SearchTagValuesV2(ctx context.Context, req *tempopb.SearchTagV
 		return nil, fmt.Errorf("error querying ingesters in Querier.SearchTagValues: %w", err)
 	}
 
-	stopEarly := false
+outerLoop:
 	for _, resp := range lookupResults {
 		for _, res := range resp.response.(*tempopb.SearchTagValuesV2Response).TagValues {
 			if distinctValues.Collect(*res) {
-				stopEarly = true
-				break
+				break outerLoop // break out of all loops
 			}
-		}
-		if stopEarly {
-			break
 		}
 	}