--- a/pkg/collector/distinct_value_collector.go
+++ b/pkg/collector/distinct_value_collector.go
@@ -5,57 +5,74 @@ import (
 )
 
 type DistinctValue[T comparable] struct {
-	values   map[T]struct{}
-	new      map[T]struct{}
-	len      func(T) int
-	maxLen   int
-	currLen  int
-	totalLen int
-	mtx      sync.RWMutex
+	values      map[T]struct{}
+	new         map[T]struct{}
+	len         func(T) int
+	maxLen      int
+	currLen     int
+	limExceeded bool
+	diffEnabled bool
+	mtx         sync.Mutex
 }
 
 // NewDistinctValue with the given maximum data size. This is calculated
 // as the total length of the recorded strings. For ease of use, maximum=0
 // is interpreted as unlimited.
+// Use NewDistinctValueWithDiff to enable diff support, but that one is slightly slower.
 func NewDistinctValue[T comparable](maxDataSize int, len func(T) int) *DistinctValue[T] {
 	return &DistinctValue[T]{
-		values: make(map[T]struct{}),
-		new:    make(map[T]struct{}),
-		maxLen: maxDataSize,
-		len:    len,
+		values:      make(map[T]struct{}),
+		new:         make(map[T]struct{}),
+		maxLen:      maxDataSize,
+		diffEnabled: false, // disable diff to make it faster
+		len:         len,
 	}
 }
 
-func (d *DistinctValue[T]) Collect(v T) (exceeded bool) {
-	d.mtx.RLock()
-	if _, ok := d.values[v]; ok {
-		d.mtx.RUnlock()
-		return // Already present
+// NewDistinctValueWithDiff is like NewDistinctValue but with diff support enabled.
+func NewDistinctValueWithDiff[T comparable](maxDataSize int, len func(T) int) *DistinctValue[T] {
+	return &DistinctValue[T]{
+		values:      make(map[T]struct{}),
+		new:         make(map[T]struct{}),
+		maxLen:      maxDataSize,
+		diffEnabled: true,
+		len:         len,
 	}
-	d.mtx.RUnlock()
+}
+
+// Collect adds a new value to the distinct value collector.
+// return true when it reaches the limits and can't fit more values.
+// callers of return of Collect or call Exceeded to stop early.
+func (d *DistinctValue[T]) Collect(v T) (exceeded bool) {
+	d.mtx.Lock()
+	defer d.mtx.Unlock()
+
+	if d.limExceeded {
+		return true
+	}
 
 	// Calculate length
 	valueLen := d.len(v)
 
-	d.mtx.Lock()
-	defer d.mtx.Unlock()
+	// Can it fit?
+	// note: we will stop adding values slightly before the limit is reached
+	if d.maxLen > 0 && d.currLen+valueLen >= d.maxLen {
+		// No, it can't fit
+		exceeded = true
+		return true
+	}
 
 	if _, ok := d.values[v]; ok {
 		return // Already present
 	}
 
-	// Record total inspected length regardless
-	d.totalLen += valueLen
-
-	// Can it fit?
-	if d.maxLen > 0 && d.currLen+valueLen > d.maxLen {
-		// No
-		return true
+	if d.diffEnabled {
+		d.new[v] = struct{}{}
 	}
 
-	d.new[v] = struct{}{}
 	d.values[v] = struct{}{}
 	d.currLen += valueLen
+
 	return false
 }
 
@@ -63,8 +80,8 @@ func (d *DistinctValue[T]) Collect(v T) (exceeded bool) {
 func (d *DistinctValue[T]) Values() []T {
 	ss := make([]T, 0, len(d.values))
 
-	d.mtx.RLock()
-	defer d.mtx.RUnlock()
+	d.mtx.Lock()
+	defer d.mtx.Unlock()
 
 	for k := range d.values {
 		ss = append(ss, k)
@@ -73,26 +90,32 @@ func (d *DistinctValue[T]) Values() []T {
 	return ss
 }
 
-// Exceeded indicates if some values were lost because the maximum size limit was met.
+// Exceeded indicates that
+// if we get rid of totalLen, then Exceeded won't work as expected
 func (d *DistinctValue[T]) Exceeded() bool {
-	d.mtx.RLock()
-	defer d.mtx.RUnlock()
-	return d.totalLen > d.currLen
+	d.mtx.Lock()
+	defer d.mtx.Unlock()
+	return d.limExceeded
 }
 
-// TotalDataSize is the total size of all distinct strings encountered.
-func (d *DistinctValue[T]) TotalDataSize() int {
-	d.mtx.RLock()
-	defer d.mtx.RUnlock()
-	return d.totalLen
+// Size is the total size of all distinct items collected
+func (d *DistinctValue[T]) Size() int {
+	d.mtx.Lock()
+	defer d.mtx.Unlock()
+	return d.currLen
 }
 
 // Diff returns all new strings collected since the last time diff was called
+// returns nil if diff is not enabled
 func (d *DistinctValue[T]) Diff() []T {
+	if !d.diffEnabled {
+		return nil
+	}
+
 	ss := make([]T, 0, len(d.new))
 
-	d.mtx.RLock()
-	defer d.mtx.RUnlock()
+	d.mtx.Lock()
+	defer d.mtx.Unlock()
 
 	for k := range d.new {
 		ss = append(ss, k)