--- a/pkg/collector/distinct_value_collector_test.go
+++ b/pkg/collector/distinct_value_collector_test.go
@@ -1,9 +1,12 @@
 package collector
 
 import (
+	"fmt"
 	"sort"
+	"strconv"
 	"testing"
 
+	"github.com/grafana/tempo/pkg/tempopb"
 	"github.com/stretchr/testify/require"
 )
 
@@ -27,4 +30,45 @@ func stringsSlicesEqual(t *testing.T, a, b []string) {
 	sort.Strings(a)
 	sort.Strings(b)
 	require.Equal(t, a, b)
+}
+
+func BenchmarkCollect(b *testing.B) {
+	// simulate 100 ingesters, each returning 10_000 tag values
+	numIngesters := 100
+	numTagValuesPerIngester := 10_000
+	ingesterTagValues := make([][]tempopb.TagValue, numIngesters)
+	for i := 0; i < numIngesters; i++ {
+		tagValues := make([]tempopb.TagValue, numTagValuesPerIngester)
+		for j := 0; j < numTagValuesPerIngester; j++ {
+			tagValues[j] = tempopb.TagValue{
+				Type:  fmt.Sprintf("string"),
+				Value: fmt.Sprintf("value_%d_%d", i, j),
+			}
+		}
+		ingesterTagValues[i] = tagValues
+	}
+
+	limits := []int{
+		0,          // no limit
+		100_000,    // 100KB
+		1_000_000,  // 1MB
+		10_000_000, // 10MB
+	}
+
+	b.ResetTimer() // to exclude the setup time for generating tag values
+	for _, lim := range limits {
+		b.Run("limit:"+strconv.Itoa(lim), func(b *testing.B) {
+			for n := 0; n < b.N; n++ {
+				// NewDistinctValue is collecting tag values without diff support
+				distinctValues := NewDistinctValue(lim, func(v tempopb.TagValue) int { return len(v.Type) + len(v.Value) })
+				for _, tagValues := range ingesterTagValues {
+					for _, v := range tagValues {
+						if distinctValues.Collect(v) {
+							break // stop early if limit is reached
+						}
+					}
+				}
+			}
+		})
+	}
 }